linuxkpi wifi multicast: purecap support.

+++ sys/compat/linuxkpi/common/include/net/mac80211.h
@@ -886,7 +886,7 @@ struct ieee80211_ops {
 	void (*sta_set_4addr)(struct ieee80211_hw *, struct ieee80211_vif *, struct ieee80211_sta *, bool);
 	void (*sta_set_decap_offload)(struct ieee80211_hw *, struct ieee80211_vif *, struct ieee80211_sta *, bool);
 
-	u64  (*prepare_multicast)(struct ieee80211_hw *, struct netdev_hw_addr_list *);
+	uintptr_t  (*prepare_multicast)(struct ieee80211_hw *, struct netdev_hw_addr_list *);
 
 	int  (*ampdu_action)(struct ieee80211_hw *, struct ieee80211_vif *, struct ieee80211_ampdu_params *);
 
@@ -915,7 +915,7 @@ struct ieee80211_ops {
 	int  (*remain_on_channel)(struct ieee80211_hw *, struct ieee80211_vif *, struct ieee80211_channel *, int, enum ieee80211_roc_type);
 	int  (*cancel_remain_on_channel)(struct ieee80211_hw *, struct ieee80211_vif *);
 
-	void (*configure_filter)(struct ieee80211_hw *, unsigned int, unsigned int *, u64);
+	void (*configure_filter)(struct ieee80211_hw *, unsigned int, unsigned int *, uintptr_t);
 	void (*config_iface_filter)(struct ieee80211_hw *, struct ieee80211_vif *, unsigned int, unsigned int);
 
 	void (*bss_info_changed)(struct ieee80211_hw *, struct ieee80211_vif *, struct ieee80211_bss_conf *, u64);
@@ -2476,7 +2476,8 @@ ieeee80211_obss_color_collision_notify(struct ieee80211_vif *vif,
 //   "updated": 20230424,
 //   "target_type": "kernel",
 //   "changes_purecap": [
-//     "subobject_bounds"
+//     "subobject_bounds",
+//     "pointer_as_integer"
 //   ]
 // }
 // CHERI CHANGES END
